[FEATURE] Move Bootstrap classes to compat folder
This patch also introduces an own class loading function to load compat
classes on demand.

Furthermore deprecated entry points for old PHPUnit configuration are
removed.
