--- src/TestingFramework/Bootstrap/UnitTestsBootstrap.php
@@ -1,17 +1,0 @@
-<?php
-/*
- * This file is part of the NIMUT testing-framework project.
- *
- * It is free software; you can redistribute it and/or modify it under
- * the terms of the GNU General Public License, either version 2
- * of the License, or any later version.
- *
- * For the full copyright and license information, please read
- * LICENSE file that was distributed with this source code.
- */
-
-/**
- * @deprecated Please use the bootstrap file /res/Configuration/UnitTestsBootstrap.php
- */
-echo 'This file will be removed with version 2.0.0. Please use the bootstrap file in /res/Configuration/FunctionalTestsBootstrap.php' . PHP_EOL;
-require __DIR__ . '/../../../res/Configuration/UnitTestsBootstrap.php';
